fix some import errors

+++ go_bindings.go
@@ -6,22 +6,15 @@ package masshortcut
 // #cgo LDFLAGS: -framework Cocoa -framework CoreFoundation -framework Carbon -framework AppKit
 // #import "./MASShortcutMonitor.h"
 // void CHotkeyCallback(int CIndex);
-// MASShortcutMonitor* get_default_shortcut_monitor() {
-//     return [MASShortcutMonitor sharedMonitor];
-// }
-// void unload_all_shortcuts(MASShortcutMonitor* monitor) {
-//     [monitor unregisterAllShortcuts];
-// }
-// void register_shortcut(MASShortcutMonitor* monitor, int Keys, int Modifiers, int HotkeyID) {
-//     MASShortcut* shortcut = [MASShortcut shortcutWithKeyCode:Keys modifierFlags:Modifiers];
-//     [monitor registerShortcut:shortcut withAction:^(void) { CHotkeyCallback(HotkeyID); }];
-// }
+// MASShortcutMonitor* get_default_shortcut_monitor();
+// void unload_all_shortcuts(MASShortcutMonitor* monitor);
+// void register_shortcut(MASShortcutMonitor* monitor, int Keys, int Modifiers, int HotkeyID);
 import "C"
 import "sync"
 
 var (
-	cbMap = map[int]func(){}
-	currentShortcut = 0
+	cbMap               = map[int]func(){}
+	currentShortcut     = 0
 	currentShortcutLock = sync.Mutex{}
 )
 
@@ -39,8 +32,8 @@ func CHotkeyCallback(CIndex C.int) {
 // ShortcutMonitor is used to define a shortcut monitor.
 type ShortcutMonitor struct {
 	monitor *C.MASShortcutMonitor
-	cbs []int
-	cbLock sync.Mutex
+	cbs     []int
+	cbLock  sync.Mutex
 }
 
 // RegisterShortcut is used to register a shortcut.
@@ -53,7 +46,7 @@ func (s *ShortcutMonitor) RegisterShortcut(Keys, Modifiers int, Callback func())
 	s.cbLock.Lock()
 	s.cbs = append(s.cbs, ThisID)
 	s.cbLock.Unlock()
-	C.register_shortcut(s.monitor, C.int(Keys), C.int(Modifiers), ThisID)
+	C.register_shortcut(s.monitor, C.int(Keys), C.int(Modifiers), C.int(ThisID))
 }
 
 // UnregisterShortcuts is used to unregister all of the shortcuts.
